feat(scopes)!: enforce scopes restrictions on a wider range of requests (#586)
BREAKING CHANGE: the public API of forest-express has changed to include the parameters which are needed to evaluate scope filters. This includes PermissionMiddlewareCreator, all classes on the /exposed folder and most services. Many routes now require providing the current user timezone.

+++ src/services/resources-remover.js
@@ -1,17 +1,39 @@
 const { InvalidParameterError } = require('./errors');
+const ResourcesGetter = require('./resources-getter');
 
 class ResourcesRemover {
-  constructor(model, ids) {
+  constructor(model, params, ids, user) {
     this._model = model;
+    this._params = params;
     this._ids = ids;
+    this._user = user;
   }
 
-  perform() {
+  async perform() {
     if (!Array.isArray(this._ids) || !this._ids.length) {
       throw new InvalidParameterError('`ids` must be a non-empty array.');
     }
 
-    return this._model.deleteMany({ _id: this._ids });
+    if (await this._isAllowed()) {
+      return this._model.deleteMany({ _id: this._ids });
+    }
+
+    return null;
+  }
+
+  /**
+   * Check if all records maching the provided list of ids are in scope.
+   * We can't do that in a single request as mongo does not supports deletes in
+   * aggregation pipelines
+   */
+  async _isAllowed() {
+    const params = {
+      timezone: this._params.timezone,
+      filters: JSON.stringify({ field: '_id', operator: 'in', value: this._ids }),
+    };
+
+    return this._ids.length
+      === await new ResourcesGetter(this._model, null, params, this._user).count();
   }
 }
